+++ src/sources/mod.rs
@@ -15,6 +15,8 @@ pub mod docker_logs;
 pub mod file;
 #[cfg(feature = "sources-generator")]
 pub mod generator;
+#[cfg(feature = "sources-heroku-logs")]
+pub mod heroku_logs;
 #[cfg(feature = "sources-host_metrics")]
 pub mod host_metrics;
 #[cfg(feature = "sources-http")]
@@ -27,8 +29,6 @@ pub mod journald;
 pub mod kafka;
 #[cfg(feature = "sources-kubernetes-logs")]
 pub mod kubernetes_logs;
-#[cfg(feature = "sources-heroku-logs")]
-pub mod logplex;
 #[cfg(feature = "sources-mongodb_metrics")]
 pub mod mongodb_metrics;
 #[cfg(feature = "sources-nginx_metrics")]
